+++ src/views/demos/DemoFileOutput.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { TextField } from '@mui/material'
 import { IDemoView } from './IDemoView'
 import { FilterRangeHandler } from '../../sonification/handler/FilterRangeHandler'
-import { FileOutput } from '../../sonification/output/FileOutputNew'
+import { FileOutput } from '../../sonification/output/FileOutput'
 import { DemoSimple, DemoSimpleProps, DemoSimpleState } from './DemoSimple'
 import { NoteHandler } from '../../sonification/handler/NoteHandler'
 import { OutputEngine } from '../../sonification/OutputEngine'
